backend/requirements.txt
Django==6.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
python-dotenv==1.2.1
Pillow==12.1.0
mysqlclient==2.2.7
PyJWT==2.8.0
bcrypt==4.1.2
djangorestframework-simplejwt==5.3.1
